requirements.txt
Pillow==3.2.0
